refactor: market-session 로직을 siglens-core로 단일화 (Phase 2)
@y0ngha/siglens-core 0.14.0이 US 정규 세션 감지를 공개하면서
siglens의 중복 로직을 제거하고 core에서 import하도록 변경.
공개된 interested staleness 휴리스틱은 유지하며
options-specific 파일로 이동.
