+++ package.json
@@ -49,7 +49,7 @@
         "@upstash/redis": "^1.37.0",
         "@vercel/analytics": "^2.0.1",
         "@vercel/functions": "^3.4.3",
-        "@y0ngha/siglens-core": "0.13.0",
+        "@y0ngha/siglens-core": "0.14.0",
         "bcryptjs": "^3.0.3",
         "clsx": "^2.1.1",
         "drizzle-orm": "^0.45.2",
