+++ src/entities/options-chain/lib/optionsCacheLife.ts
@@ -1,4 +1,4 @@
-import { getEtSessionStatus } from '@/shared/lib/marketSession';
+import { getEtSessionStatus } from '@y0ngha/siglens-core';
 
 export type OptionsCacheLifeProfile =
     | 'options-market-open'
@@ -10,8 +10,8 @@ export type OptionsCacheLifeProfile =
  * registered in `next.config.ts`'s `cacheLife` map.
  *
  * - Saturday/Sunday → `options-weekend`
- * - Weekday 09:30~16:00 ET → `options-market-open`
- * - Weekday off-hours → `options-market-closed`
+ * - Weekday 09:30–16:00 ET (16:00 close exclusive) → `options-market-open`
+ * - Weekday off-hours (incl. 16:00 sharp) → `options-market-closed`
  */
 export function getOptionsCacheLifeProfile(
     now: Date = new Date()
